+++ packages/github/package.json
@@ -40,6 +40,7 @@
     "@actions/http-client": "^1.0.3",
     "@octokit/core": "^2.5.0",
     "@octokit/graphql": "^4.3.1",
+    "@octokit/plugin-enterprise-server": "^7.4.1",
     "@octokit/plugin-paginate-rest": "^2.2.0",
     "@octokit/plugin-rest-endpoint-methods": "^3.10.0"
   },
